main/requirements.txt
boto3==1.35.40
confluent_kafka==2.5.3
kafka_python==2.0.2
